third_party/WebKit/LayoutTests/fast/css/getComputedStyle/getComputedStyle-border-radius-shorthand-expected.txt

Tests that the border-radius shorthand is computed properly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS computedStyle.getPropertyValue('border-radius') is '64px'
PASS computedStyle.getPropertyValue('border-radius') is '32px 16px 64px / 8px 48px'
PASS computedStyle.getPropertyValue('border-radius') is '64px'
PASS computedStyle.getPropertyValue('border-radius') is '1px 2px'
PASS computedStyle.getPropertyValue('border-radius') is '1px 2px 4px'
PASS computedStyle.getPropertyValue('border-radius') is '1px 2px 4px 6px'
PASS computedStyle.getPropertyValue('border-radius') is '1px 2px 4px 6px / 64px'
PASS computedStyle.getPropertyValue('border-radius') is '1px 2px 4px 6px / 64px 3px'
PASS computedStyle.getPropertyValue('border-radius') is '1px 2px 4px 6px / 64px 3px 7px'
PASS computedStyle.getPropertyValue('border-radius') is '1px 2px 4px 6px / 64px 3px 7px 8px'
PASS computedStyle.getPropertyValue('border-radius') is '10px 20px 30px 40px / 15px 25px 35px 45px'
PASS computedStyle.getPropertyValue('border-radius') is '10px 20px 30px 40px / 10px 25px 35px 45px'
PASS computedStyle.getPropertyValue('border-radius') is '10px 20px 30px 40px / 10px 20px 35px 45px'
PASS computedStyle.getPropertyValue('border-radius') is '10px 20px 30px 40px / 10px 20px 30px 45px'
PASS successfullyParsed is true

TEST COMPLETE
